oauth-proxy: Added 'signature-host' property.
This enables signing with a different hostname than the one in the endpoint.
The case in point is Photobucket: It has "silo" URLs that are used for
certain types of operations, such as api123.photobucket.com, yet the signature
still expects to use api.photobucket.com in the base string.

Fixes: https://bugs.meego.com/show_bug.cgi?id=14192