Avoid infinite loop with wrong HTTP credentials
When provided with wrong credentials, libsoup will try to connect
and emit its 'authenticate' signal as long as its callback calls
soup_auth_authenticate. It will fail the request and report to
the caller if this function is not called. Since the 'retrying'
parameter to the 'authenticate' callback lets us know when the
credentials we provided are the wrong ones, this commit makes sure
we stop calling soup_auth_authenticate after trying the credentials
once. Without this, libsoup will try the same request again and
again without ever returning when provided with wrong credentials.

https://bugzilla.gnome.org/show_bug.cgi?id=658937