rest-proxy-call: Include the zero byte added by soup to the payload
In the previous commit the downloading of binary data was fixed -
unfortunately this broke the behaviour of downloading textual data where
the clients expect a NUL terminated string for the payload.

This change makes sure to include the zero byte that soup adds to
support this use case and as a result the test suite passes.