Allow to modify REST function in serialize_params vfunc
RestProxyCall::prepare_message() calls the serialize_params vfunc
so that child classes can serialize the call parameters in whichever
way they want.
One way of doing that could be to append the parameters to the URI
that is called (http://example.com?param1=value1;param2=value2).
However, the URI to call is determined at the beginning of
prepare_message(), and is not refreshed after calling
RestProxyCall::serialize_params(), so it's not possible to
append parameters to the URI that is going to be called.

This commit rebuilds the URI to call after calling serialize_params()
in case it has changed.

https://bugzilla.gnome.org/show_bug.cgi?id=708359