oauth: Make sure RestProxyCall::url is set before using it
The oauth "prepare" step needs RestProxyCall::url to be set, but since
commit c66b6d this is only set after the "prepare" and
"serialize_headers" methods have been called. As it's
better if OAuthProxyCall do not directly access RestProxyCall private
data, this commit adds a getter to RestProxyCall which will be used
to make sure RestProxyCall::url is set when OAuthProxyCall::prepare
needs it.

https://bugzilla.gnome.org/show_bug.cgi?id=708359