lastfm: Fix function setting regression
Since commit c66b6d, RestProxyCall::url is regenerated after calling the
RestProxyCall::prepare virtual method. This breaks the lastfm code as
it needs to reset RestProxyCall::url in order to remove the function
from it.

It used to do that by directly accessing RestProxyCall private
data. Since c66b6d, this can be solved using only public methods as
if the function is reset on the RestProxyCall in ::prepare, ::url will
be regenerated to reflect that after ::prepare and ::serialize_params
have been called.

https://bugzilla.gnome.org/show_bug.cgi?id=708359