oauth: Don't leak temp data in steal_oauth_params
steal_oauth_params() creates a list of strdup'ed strings containing the
name of the ouath parameters which should be stolen from the call
arguments. However, the code makes sure to free the list, but never
frees its content, causing a leak. This commit makes sure the dup'ed
strings are freed.