build-sys: Set version to 0.8.0
We set rest_version to 0.8.0 in configure.ac, but we cannot define it in
terms of api_major, api_minor and api_micro as setting
api_major.api_minor to 0.8 means an ABI break.
Instead, rest package version and api_version are decoupled, api_version
can stay at 0.7 while the package version is set to 0.8.0.