build: Set API version to 1.0
This will change the installed library name from librest-0.7.so to
librest-1.0.so as well as the name of the installed .pc file. In short,
API and ABI are going to be broken. As the subsequent commits are going
to remove deprecated symbols, and change the public API, this is
actually desired.
Note that older librest versions should be parallel installable with the
newer ones.