xml-node: Define the order in which attributes & children are printed
The order in which GHashTable returns its key-value pairs is undefined.
Therefore the output of rest_xml_node_print can change based on the
GHashTable implementation. While not strictly necessary, it would be
nice to avoid that. Having a stable order, even if it is not
documented and depends on the current RestXmlNode code, is handy for
testing.

This was the main reason behind the tests/xml.c breakage.

https://bugzilla.gnome.org/show_bug.cgi?id=788960