Fix the declaration of the RestOAuth2Error quark function
The missing `void` breaks the strict prototypes warning.

In C, a function with no arguments and a function with `void` as an
argument are fundamentally different; the former can be called with any
argument, while the latter can only be called with no arguments.

Since rest_oauth2_error_quark() is defined with `void` by the
G_DEFINE_QUARK macro, its declaration should be made to match.